Solved Socket.io connection issue
Issue created by script trying to update slider elements that were not
shown on the page_home. Added an if statement so that they are only
updated if the user is on page_settings and also a refresh that is run
during page changes to ensure it is always showing the latest info.